GH-406: Use headless extenion registrion helper everywhere
