+++ plugins/org.eclipse.n4js.generator.headless/META-INF/MANIFEST.MF
@@ -9,6 +9,10 @@ Require-Bundle: org.eclipse.xtext.generator,
  org.eclipse.core.resources,
  org.eclipse.n4js,
  org.eclipse.n4js.utils,
- org.eclipse.n4js.transpiler.es
+ org.eclipse.n4js.transpiler.es,
+ org.eclipse.n4js.runner,
+ org.eclipse.n4js.tester,
+ org.eclipse.n4js.runner.nodejs,
+ org.eclipse.n4js.tester.nodejs
 Export-Package: org.eclipse.n4js.generator.headless,
  org.eclipse.n4js.generator.headless.logging
