+++ plugins/org.eclipse.n4js.runner/src/org/eclipse/n4js/runner/extension/RunnerRegistry.java
@@ -18,13 +18,12 @@ import org.apache.log4j.Logger;
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.IExtensionRegistry;
 import org.eclipse.core.runtime.RegistryFactory;
+import org.eclipse.n4js.runner.IRunner;
+import org.eclipse.n4js.runner.RunConfiguration;
 
 import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Singleton;
-
-import org.eclipse.n4js.runner.IRunner;
-import org.eclipse.n4js.runner.RunConfiguration;
 
 /**
  * Utility for working with runner description extension point.
@@ -52,9 +51,10 @@ public class RunnerRegistry {
 	 */
 	public void register(IRunnerDescriptor runnerDescriptor) {
 		final String runnerId = runnerDescriptor.getId();
-		if (descriptors.containsKey(runnerId))
-			throw new IllegalArgumentException("cannot register two runners with the same ID: " + runnerId);
-		descriptors.put(runnerId, runnerDescriptor);
+		if (!descriptors.containsKey(runnerId)) {
+			descriptors.put(runnerId, runnerDescriptor);
+		}
+		// TODO: throw new IllegalArgumentException("cannot register two runners with the same ID: " + runnerId);
 	}
 
 	/**
