+++ plugins/org.eclipse.n4js.tester/src/org/eclipse/n4js/tester/extension/TesterRegistry.java
@@ -18,13 +18,12 @@ import org.apache.log4j.Logger;
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.IExtensionRegistry;
 import org.eclipse.core.runtime.RegistryFactory;
+import org.eclipse.n4js.tester.ITester;
+import org.eclipse.n4js.tester.TestConfiguration;
 
 import com.google.inject.Inject;
 import com.google.inject.Injector;
 import com.google.inject.Singleton;
-
-import org.eclipse.n4js.tester.ITester;
-import org.eclipse.n4js.tester.TestConfiguration;
 
 /**
  * Utility for working with the 'testers' extension point.
@@ -52,9 +51,10 @@ public class TesterRegistry {
 	 */
 	public void register(ITesterDescriptor testerDescriptor) {
 		final String testerId = testerDescriptor.getId();
-		if (descriptors.containsKey(testerId))
-			throw new IllegalArgumentException("cannot register two testers with the same ID: " + testerId);
-		descriptors.put(testerId, testerDescriptor);
+		if (!descriptors.containsKey(testerId)) {
+			descriptors.put(testerId, testerDescriptor);
+		}
+		// throw new IllegalArgumentException("cannot register two testers with the same ID: " + testerId);
 	}
 
 	/**
